:recycle: Refactor: Composable 분리
